moved test around
