+++ dgraph/cmd/live/load-uids/xid.schema
