--- dgraph/cmd/live/load-upsert/load_test.go
@@ -1,120 +1,0 @@
-/*
- * Copyright 2017-2018 Dgraph Labs, Inc. and Contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package main
-
-import (
-	"context"
-	"io/ioutil"
-	"log"
-	"os"
-	"path"
-	"runtime"
-	"testing"
-
-	"github.com/dgraph-io/dgo/v200"
-	"github.com/dgraph-io/dgo/v200/protos/api"
-	"github.com/stretchr/testify/require"
-
-	"github.com/dgraph-io/dgraph/testutil"
-	"github.com/dgraph-io/dgraph/x"
-)
-
-var alphaService = testutil.SockAddr
-var zeroService = testutil.SockAddrZero
-
-var (
-	testDataDir string
-	dg          *dgo.Dgraph
-)
-
-const (
-	alphaName = "alpha1"
-)
-
-func checkLoadedData(t *testing.T, newUids bool) {
-	resp, err := dg.NewTxn().Query(context.Background(), `
-		{
-			q(func: eq(xid, "m.1234")) {
-				xid
-				name
-				value
-			}
-		}
-	`)
-	require.NoError(t, err)
-
-	gotMap := testutil.UnmarshalJSON(t, string(resp.GetJson()))
-	wantMap := testutil.UnmarshalJSON(t, `
-		{
-		    "q": [
-			    {
-					"xid": "m.1234",
-					"name": "name 1234",
-					"value": "value 1234"
-			    }
-			]
-		}
-	`)
-
-	testutil.CompareJSONMaps(t, wantMap, gotMap)
-}
-
-func TestLiveLoadUpsert(t *testing.T) {
-	testutil.DropAll(t, dg)
-
-	pipeline := [][]string{
-		{testutil.DgraphBinaryPath(), "live",
-			"--schema", testDataDir + "/xid.schema", "--files", testDataDir + "/xid_a.rdf",
-			"--alpha", alphaService, "--zero", zeroService, "-u", "groot", "-p", "password",
-			"-U", "xid"},
-	}
-	_, err := testutil.Pipeline(pipeline)
-	require.NoError(t, err, "live loading JSON file exited with error")
-
-	pipeline = [][]string{
-		{testutil.DgraphBinaryPath(), "live",
-			"--schema", testDataDir + "/xid.schema", "--files", testDataDir + "/xid_b.rdf",
-			"--alpha", alphaService, "--zero", zeroService, "-u", "groot", "-p", "password",
-			"-U", "xid"},
-	}
-	_, err = testutil.Pipeline(pipeline)
-	require.NoError(t, err, "live loading JSON file exited with error")
-
-	checkLoadedData(t, false)
-}
-
-func TestMain(m *testing.M) {
-	_, thisFile, _, _ := runtime.Caller(0)
-	testDataDir = path.Dir(thisFile)
-
-	var err error
-	dg, err = testutil.DgraphClientWithGroot(testutil.SockAddr)
-	if err != nil {
-		log.Fatalf("Error while getting a dgraph client: %v", err)
-	}
-	x.Check(dg.Alter(
-		context.Background(), &api.Operation{DropAll: true}))
-
-	// Try to create any files in a dedicated temp directory that gets cleaned up
-	// instead of all over /tmp or the working directory.
-	tmpDir, err := ioutil.TempDir("", "test.tmp-")
-	x.Check(err)
-	os.Chdir(tmpDir)
-	defer os.RemoveAll(tmpDir)
-
-	os.Exit(m.Run())
-}
